extract command model from uploader to core
