+++ audit/model_signals.py
@@ -10,13 +10,12 @@ from audit.models import CofkUnionAuditLiteral
 from core import constant, recref_settings
 from core.helper import model_utils
 from core.models import CofkUnionComment, CofkUnionRelationship, CofkUnionRelationshipType, CofkUnionResource, Recref, \
-    CofkUnionNationality
+    CofkUnionNationality, CofkUnionImage, CofkUnionRoleCategory, CofkUnionSubject
 from institution.models import CofkUnionInstitution
 from location.models import CofkUnionLocation
 from manifestation.models import CofkUnionManifestation
 from person.models import CofkUnionPerson
 from publication.models import CofkUnionPublication
-from uploader.models import CofkUnionImage, CofkUnionSubject, CofkUnionRoleCategory
 from work.models import CofkUnionWork
 
 log = logging.getLogger(__name__)
