+++ core/helper/recref_handler.py
@@ -23,9 +23,8 @@ from core.forms import ImageForm, UploadImageForm, ResourceForm, RecrefForm
 from core.helper import recref_utils, model_utils, iter_utils
 from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.view_utils import create_formset
-from core.models import Recref
+from core.models import Recref, CofkUnionImage, CofkUnionRoleCategory, CofkUnionSubject
 from core.services import media_service
-from uploader.models import CofkUnionSubject, CofkUnionRoleCategory, CofkUnionImage
 
 log = logging.getLogger(__name__)
