+++ location/models.py
@@ -59,7 +59,7 @@ class CofkUnionLocation(models.Model, RecordTracker):
     element_7_eg_empire = models.CharField(max_length=100)
     uuid = models.UUIDField(blank=True, null=True)
 
-    images = models.ManyToManyField('uploader.CofkUnionImage')  # TOBEREMOVE
+    images = models.ManyToManyField('core.CofkUnionImage')  # TOBEREMOVE
 
     @property
     def comments(self):
@@ -107,7 +107,7 @@ class CofkLocationResourceMap(Recref):
 
 class CofkLocationImageMap(Recref):
     location = models.ForeignKey(CofkUnionLocation, on_delete=models.CASCADE)
-    image = models.ForeignKey('uploader.CofkUnionImage', on_delete=models.CASCADE)
+    image = models.ForeignKey('core.CofkUnionImage', on_delete=models.CASCADE)
 
     class Meta(Recref.Meta):
         db_table = 'cofk_location_image_map'
