+++ uploader/entities/work.py
@@ -8,7 +8,8 @@ from django.core.exceptions import ValidationError
 from uploader.entities.entity import CofkEntity
 from uploader.entities.locations import CofkLocations
 from uploader.entities.people import CofkPeople
-from uploader.models import CofkCollectUpload, CofkCollectStatus, Iso639LanguageCode
+from uploader.models import CofkCollectUpload, CofkCollectStatus
+from core.models import Iso639LanguageCode
 from work.models import CofkCollectWork, CofkCollectLanguageOfWork, CofkCollectWorkResource, \
     CofkCollectPersonMentionedInWork, CofkCollectAuthorOfWork, CofkCollectAddresseeOfWork, CofkCollectOriginOfWork, \
     CofkCollectDestinationOfWork
